requirements.txt
flask
testcontainers
psycopg2-binary
paho-mqtt
Docker